+++ src/feature/scrap/hook/index.js
@@ -1,0 +1,1 @@
+export { default as useScrap } from './useScrap';
