+++ src/hooks/index.js
@@ -1,2 +1,1 @@
-export { default as useScrap } from './useScrap';
 export { default as useSearch } from './useSearch';
